Add_Animation can now be used to add plain images.
If Add_Animation is called with a filename that has an extension ".png",
it will add the single image in the normal manor and create a named
entry for it.  Set_Named_Animation and Get_Named_Animation_Range can
still be used to set the animation and determine the range.  This works
since Set_Named_Animation will disable the animation if there is only
one image.