Background now uses image sets.
The background manager uses cSprite in order to take advantage of its
animation code, but that is all.  It then directly accesses the images
in cSprite to use for the background.  In order to prevent the level
from doing anything with sprites, a custom sprite manager is used for
the cSprites in a background.