Refactored image animation code to external class.
Image animation code is now handled by cImageSet.  This class is meant
to be used as a base class, and requires the derived class to provide
certain virtual functions:

* std::string Get_Identity(void) - optional
  This should return a string identifying the object.  It is used in
  logging messages by the image set.
* void Set_Image_Set_Image(cGL_Surface* image, bool new_startimage = 0)
  This is required and should set the active image to the specifed
  image.

Set_Image_Num no longer takes an option to delete the image when changing
away from it.  The reason is if a call to Set_Image deletes the
cGL_Surface, the image set will still hold a dangling pointer to it.
However, direct calls to Set_Image can still specify to delete when
chaning.

One option to allow the above may be to provide an option to Add_Image to
allow deletion of the images, not when switching away from the image via
Set_Image, but when the image set is destroyed. Add_Image_Set will have
no such option because it uses pVideo to load the surface, which seems
to manage them externally.